feat(spec): absorb retrieval protocols, prune engine/retrieval/ infra
Soul Protocol is a standard. Concrete retrieval orchestration (router,
broker impl, callable-wrapping projection adapter) is application layer
and belongs in the consuming runtime, not the spec. This commit moves
the vocabulary a third-party implementation needs into spec/retrieval.py
and deletes the infra that used to live under engine/retrieval/.

Moved into spec/retrieval.py:
  - Credential (BaseModel) — the short-lived token shape
  - CredentialBroker (Protocol) — four-method interface
  - SourceAdapter (Protocol) — sync retrieval interface
  - AsyncSourceAdapter (Protocol) — structural tag for async-capable adapters
  - RetrievalError hierarchy — NoSourcesError, SourceTimeoutError,
    CredentialScopeError, CredentialExpiredError

Deleted from soul-protocol:
  - engine/retrieval/__init__.py, adapters.py, broker.py, exceptions.py,
    router.py — concrete implementations move to pocketpaw
  - tests/test_engine/test_retrieval.py — infra tests migrate with the code

Public surface updates:
  - spec/__init__.py re-exports the new protocol + exception names.
  - DataRef is re-exported as RetrievalDataRef to keep it distinct from
    the journal-layer DataRef in spec.journal (different granularity,
    same concept).

Net diff: -1501 LOC (1695 deleted, 194 added). All 2333 soul-protocol
tests pass on the thinner surface. engine/ now holds only the journal
substrate — the one engine piece that is genuinely spec-layer
(append-only log, hash chain, UTC timestamping, scope stamping).

Ships with 0.3.2. Pocketpaw receives the router + InMemoryCredentialBroker
+ ProjectionAdapter in a companion PR; both can be path-sourced against
this branch until 0.3.2 lands on PyPI.
