+++ src/soul_protocol/spec/__init__.py
@@ -14,6 +14,14 @@
 #   Workstream D.
 # Updated: feat/retrieval-trace-spec — Exported RetrievalTrace + TraceCandidate
 #   from spec.trace (the per-recall receipt model — PR #161).
+# Updated: feat/0.3.2-prune-retrieval-infra — Exported the retrieval vocabulary
+#   that used to live under engine/retrieval/ as public spec types: protocols
+#   (SourceAdapter, AsyncSourceAdapter, CredentialBroker), data classes
+#   (Credential), and the RetrievalError exception hierarchy. Concrete
+#   implementations (RetrievalRouter, InMemoryCredentialBroker,
+#   ProjectionAdapter) are application-layer and now live in pocketpaw.
+#   DataRef is re-exported as RetrievalDataRef to distinguish it from the
+#   journal-layer DataRef in spec.journal.
 
 from __future__ import annotations
 
@@ -58,6 +66,23 @@ from .memory import (
     MemoryStore,
     MemoryVisibility,
     Participant,
+)
+from .retrieval import (
+    AsyncSourceAdapter,
+    CandidateSource,
+    Credential,
+    CredentialBroker,
+    CredentialExpiredError,
+    CredentialScopeError,
+    DataRef as RetrievalDataRef,
+    NoSourcesError,
+    PointInTimeNotSupported,
+    RetrievalCandidate,
+    RetrievalError,
+    RetrievalRequest,
+    RetrievalResult,
+    SourceAdapter,
+    SourceTimeoutError,
 )
 from .scope import match_scope, normalise_scopes
 from .soul_file import pack_soul, unpack_soul, unpack_to_container
@@ -111,6 +136,25 @@ __all__ = [
     # Retrieval trace (per-recall receipt)
     "RetrievalTrace",
     "TraceCandidate",
+    # Retrieval spec (types + protocols + exceptions).
+    # Concrete implementations (RetrievalRouter, InMemoryCredentialBroker,
+    # ProjectionAdapter) live in the consuming runtime — not part of the
+    # headless standard. See ADR-NNN / docs/SPEC.md.
+    "CandidateSource",
+    "RetrievalRequest",
+    "RetrievalCandidate",
+    "RetrievalResult",
+    "RetrievalDataRef",
+    "PointInTimeNotSupported",
+    "Credential",
+    "CredentialBroker",
+    "SourceAdapter",
+    "AsyncSourceAdapter",
+    "RetrievalError",
+    "NoSourcesError",
+    "SourceTimeoutError",
+    "CredentialScopeError",
+    "CredentialExpiredError",
     # Learning
     "LearningEvent",
     "SoulTemplate",
